Refactor code, improve comments and move functions to a separate file
